modifico nomenclatura de boton-inicio a boton-inscripcion, agrego formulario de inscripcion
